+++ packages/atomic/src/components/search/facets/atomic-category-facet/atomic-category-facet.tsx
@@ -12,7 +12,7 @@ import {
   FacetConditionsManager,
 } from '@coveo/headless';
 import {Component, h, State, Prop, Element, Fragment} from '@stencil/core';
-import LeftArrow from 'coveo-styleguide/resources/icons/svg/arrow-left-rounded.svg';
+import LeftArrow from '../../../../images/arrow-left-rounded.svg';
 import {
   FocusTarget,
   FocusTargetController,